requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.1
pydantic==2.5.2
gunicorn==21.2.0

